using uint function to convert objecthandle to uint

+++ handlers/handle_generate_aes_key.go
@@ -67,7 +67,7 @@ func postGenerateAESKey(mgr *pkcs11mgr.Manager, w http.ResponseWriter, r *http.R
 	logger.AppLog.Infof("AES key generated successfully - Handle: %d", handle)
 
 	resp := models.GenAESKeyResponse{
-		Handle: handle,
+		Handle: uint(handle),
 		Label:  &req.Label,
 		Id:     &req.Id,
 		Bits:   &req.Bits,
